myapp/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
Django==3.0.1
django-filter==2.2.0
djangorestframework==3.11.0
docutils==0.16
idna==2.9
Kivy==1.11.1
kivy-deps.angle==0.2.0
kivy-deps.glew==0.2.0
kivy-deps.gstreamer==0.2.0
kivy-deps.sdl2==0.2.0
Kivy-Garden==0.1.4
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.1.2
mysqlclient==1.4.6
numpy==1.18.0
pygame==1.9.6
Pygments==2.6.1
pyparsing==2.4.6
pypiwin32==223
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
requests==2.23.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.8
